Add crypto-js to deps and mailchimp register func
Bind the function that handles submit to an api endpoint
The api endpoint is in charge of making an http post request
to the mailchimp api and register an email. The md5 hash of the email
is a required parameter. Also environmental variables are added as
values for the mailchimp settings